Solver algorithm

+++ src/game24/util/Convert.java
@@ -2,46 +2,46 @@ package game24.util;
 
 public class Convert{
     static String[] allCards = {"A", "2", "3", "4", "5", "6", "7", "8", "9", "10", "J", "Q", "K"};
-    static int[] tempValue4 = new int[4];
+    static float[] tempValue4 = new float[4];
 
     /* Validasi/konversi input */
-    public static int[] card4ToValue4(String[] cards){
+    public static float[] card4ToValue4(String[] cards){
         // Kembalikan null jika panjang array bukan 4
         if(cards.length != 4)return null;
         
         for(int i = 0; i < 4; i++){
-            int val = cardToValue(cards[i]);
+            float val = cardToValue(cards[i]);
 
             // Break, kembalikan null jika tidak valid
-            if(val == -1)return null;
+            if(val == -1f)return null;
             tempValue4[i] = val;
         }
 
         return tempValue4;
     }
 
-    public static int cardToValue(String card){
+    public static float cardToValue(String card){
         return switch(card){
-            case "A" -> 1;
-            case "2" -> 2;
-            case "3" -> 3;
-            case "4" -> 4;
-            case "5" -> 5;
-            case "6" -> 6;
-            case "7" -> 7;
-            case "8" -> 8;
-            case "9" -> 9;
-            case "10" -> 10;
-            case "J" -> 11;
-            case "Q" -> 12;
-            case "K" -> 13;
-            default -> -1;
+            case "A" -> 1f;
+            case "2" -> 2f;
+            case "3" -> 3f;
+            case "4" -> 4f;
+            case "5" -> 5f;
+            case "6" -> 6f;
+            case "7" -> 7f;
+            case "8" -> 8f;
+            case "9" -> 9f;
+            case "10" -> 10f;
+            case "J" -> 11f;
+            case "Q" -> 12f;
+            case "K" -> 13f;
+            default -> -1f;
         };
     }
 
-    public static String valueToCard(int val){
+    public static String valueToCard(float val){
         // Kembalikan null jika tidak valid
-        if(val < 1 || val > 13)return null;
-        return allCards[val - 1];
+        if(val < 1f || val > 13f)return null;
+        return allCards[(int)val - 1];
     }
 }
